refactor some scrutinizer errors
