+++ src/main/java/edu/itu/cavabunga/core/entity/Participant.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Entity
 @EntityListeners(AuditingEntityListener.class)
-public class Participant implements java.io.Serializable{
+public class Participant {
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "org.hibernate.id.UUIDGenerator")
@@ -28,7 +28,7 @@ public class Participant implements java.io.Serializable{
 
     @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL, mappedBy = "owner")
     @JsonManagedReference(value = "participantAndComponent")
-    private List<Component> components = new ArrayList<Component>();
+    private List<Component> components = new ArrayList<>();
 
     public Participant(){
